refactor: Use Transformable base class for both Table and QueryBuilder.
